+++ apps/carp_mobile_sensing_app/lib/main.dart
@@ -17,7 +17,6 @@ import 'package:carp_communication_package/communication.dart';
 import 'package:carp_apps_package/apps.dart';
 import 'package:carp_movisens_package/carp_movisens_package.dart';
 import 'package:carp_health_package/health_package.dart';
-import 'package:health/health.dart';
 import 'package:carp_movesense_package/carp_movesense_package.dart';
 // import 'package:carp_cortrium_package/carp_cortrium_package.dart';
 // import 'package:movisens_flutter/movisens_flutter.dart';